// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};
// ...
export type Category = DatabaseIdentifier & HierarchicalNode & HierarchicalTermNode & MenuItemLinkable & Node & TermNode & UniformResourceIdentifiable & {
  __typename?: 'Category';
  /** The ancestors of the node. Default ordered as lowest (closest to the child) to highest (closest to the root). */
  ancestors?: Maybe<CategoryToAncestorsCategoryConnection>;
  /**
   * The id field matches the WP_Post-&gt;ID field.
   * @deprecated Deprecated in favor of databaseId
   */
  categoryId?: Maybe<Scalars['Int']>;
  /** Connection between the Category type and the CentraldeDecorado type */
  centraldeDecorados?: Maybe<CategoryToCentraldeDecoradoConnection>;
  /** Connection between the category type and its children categories. */
  children?: Maybe<CategoryToCategoryConnection>;
  /** Connection between the Category type and the ContentNode type */
  contentNodes?: Maybe<CategoryToContentNodeConnection>;
  /** The number of objects connected to the object */
  count?: Maybe<Scalars['Int']>;
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
  /** The description of the object */
  description?: Maybe<Scalars['String']>;
  /** Connection between the Category type and the Empreendimento type */
  empreendimentos?: Maybe<CategoryToEmpreendimentoConnection>;
  /** Connection between the TermNode type and the EnqueuedScript type */
  enqueuedScripts?: Maybe<TermNodeToEnqueuedScriptConnection>;
  /** Connection between the TermNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets?: Maybe<TermNodeToEnqueuedStylesheetConnection>;
  /** The globally unique ID for the object */
  id: Scalars['ID'];
  /** Whether the node is a Comment */
  isComment: Scalars['Boolean'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node represents the front page. */
  isFrontPage: Scalars['Boolean'];
  /** Whether  the node represents the blog page. */
  isPostsPage: Scalars['Boolean'];
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The link to the term */
  link?: Maybe<Scalars['String']>;
  /** The human friendly name of the object. */
  name?: Maybe<Scalars['String']>;
  /** Connection between the category type and its parent category. */
  parent?: Maybe<CategoryToParentCategoryConnectionEdge>;
  /** Database id of the parent node */
  parentDatabaseId?: Maybe<Scalars['Int']>;
  /** The globally unique identifier of the parent node. */
  parentId?: Maybe<Scalars['ID']>;
  /** Connection between the Category type and the post type */
  posts?: Maybe<CategoryToPostConnection>;
  /** An alphanumeric identifier for the object unique to its type. */
  slug?: Maybe<Scalars['String']>;
  /** Connection between the Category type and the Taxonomy type */
  taxonomy?: Maybe<CategoryToTaxonomyConnectionEdge>;
  /** The name of the taxonomy that the object is associated with */
  taxonomyName?: Maybe<Scalars['String']>;
  /** The ID of the term group that this term object belongs to */
  termGroupId?: Maybe<Scalars['Int']>;
  /** The taxonomy ID that the object is associated with */
  termTaxonomyId?: Maybe<Scalars['Int']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};
// ...
export type CategoryToAncestorsCategoryConnection = CategoryConnection & Connection & {
  __typename?: 'CategoryToAncestorsCategoryConnection';
  /** Edges for the CategoryToAncestorsCategoryConnection connection */
  edges: Array<CategoryToAncestorsCategoryConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Category>;
  /** Information about pagination in a connection. */
  pageInfo: CategoryToAncestorsCategoryConnectionPageInfo;
};
// ...
export type DatabaseIdentifier = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
};
// ...
export type Empreendimento_Empreendimento = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento';
  andamentoDaObra?: Maybe<Array<Maybe<Empreendimento_Empreendimento_AndamentoDaObra>>>;
  arquivoDaApresentacao?: Maybe<MediaItem>;
  diferenciaisItems?: Maybe<Array<Maybe<Empreendimento_Empreendimento_DiferenciaisItems>>>;
  empCidade?: Maybe<Scalars['String']>;
  empDataDeInsercao?: Maybe<Scalars['String']>;
  empDescricao?: Maybe<Scalars['String']>;
  empDormitorios?: Maybe<Scalars['String']>;
  empMetragem?: Maybe<Scalars['String']>;
  empPrevisaoDeTermino?: Maybe<Scalars['String']>;
  empVagasDeGaragem?: Maybe<Scalars['String']>;
  endStandVendas?: Maybe<Scalars['String']>;
  enderecoBairro?: Maybe<Scalars['String']>;
  enderecoNumero?: Maybe<Scalars['String']>;
  enderecoRua?: Maybe<Scalars['String']>;
  estagioDaObra?: Maybe<Category>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  galeriaDeFotos?: Maybe<Array<Maybe<MediaItem>>>;
  imagemPrincipal?: Maybe<MediaItem>;
  imagensOutros?: Maybe<Array<Maybe<MediaItem>>>;
  itemsPlantas?: Maybe<Array<Maybe<Empreendimento_Empreendimento_ItemsPlantas>>>;
  itensAreacomuns?: Maybe<Array<Maybe<Empreendimento_Empreendimento_ItensAreacomuns>>>;
  linkDoTourVirtual?: Maybe<Scalars['String']>;
  /** Caso não tiver deixar vazio. */
  linkDoTourVirtual2?: Maybe<Scalars['String']>;
  listaItensTec?: Maybe<Scalars['String']>;
  listaItensTec2?: Maybe<Scalars['String']>;
  logotipoDoEmpreendimento?: Maybe<MediaItem>;
  nomeDoEmpreendimento?: Maybe<Scalars['String']>;
  pontosDeReferencia?: Maybe<Array<Maybe<Empreendimento_Empreendimento_PontosDeReferencia>>>;
  textoLegal?: Maybe<Scalars['String']>;
  tipoDoEmpreendimento?: Maybe<Scalars['String']>;
  totalDaObra?: Maybe<Scalars['String']>;
  videosOutos?: Maybe<Array<Maybe<Empreendimento_Empreendimento_VideosOutos>>>;
};

/** Field Group */
export type Empreendimento_Empreendimento_AndamentoDaObra = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_andamentoDaObra';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  nomeAndamentoDaObra?: Maybe<Scalars['String']>;
  valorAndamentoDaObra?: Maybe<Scalars['Float']>;
};

/** Field Group */
export type Empreendimento_Empreendimento_DiferenciaisItems = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_diferenciaisItems';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  iconeimagemDoDiferencial?: Maybe<MediaItem>;
  nomeDiferencial?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Empreendimento_Empreendimento_ItemsPlantas = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_itemsPlantas';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  imagensDasPlantas?: Maybe<MediaItem>;
  plantasDescricao?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Empreendimento_Empreendimento_ItensAreacomuns = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_itensAreacomuns';
  descricaoAreacomuns?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  iconeimagemAreacomuns?: Maybe<MediaItem>;
};

/** Field Group */
export type Empreendimento_Empreendimento_PontosDeReferencia = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_pontosDeReferencia';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  iconeOuImagemRef?: Maybe<MediaItem>;
  nomePontoReferencia?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Empreendimento_Empreendimento_VideosOutos = AcfFieldGroup & {
  __typename?: 'Empreendimento_Empreendimento_videosOutos';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  linkDoVideo?: Maybe<Scalars['String']>;
};
// ...
export type HierarchicalNode = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
  /** The globally unique ID for the object */
  id: Scalars['ID'];
  /** Database id of the parent node */
  parentDatabaseId?: Maybe<Scalars['Int']>;
  /** The globally unique identifier of the parent node. */
  parentId?: Maybe<Scalars['ID']>;
};

/** Term node with hierarchical (parent/child) relationships */
export type HierarchicalTermNode = {
  /** The number of objects connected to the object */
  count?: Maybe<Scalars['Int']>;
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
  /** The description of the object */
  description?: Maybe<Scalars['String']>;
  /** Connection between the TermNode type and the EnqueuedScript type */
  enqueuedScripts?: Maybe<TermNodeToEnqueuedScriptConnection>;
  /** Connection between the TermNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets?: Maybe<TermNodeToEnqueuedStylesheetConnection>;
  /** The globally unique ID for the object */
  id: Scalars['ID'];
  /** Whether the node is a Comment */
  isComment: Scalars['Boolean'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node represents the front page. */
  isFrontPage: Scalars['Boolean'];
  /** Whether  the node represents the blog page. */
  isPostsPage: Scalars['Boolean'];
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The link to the term */
  link?: Maybe<Scalars['String']>;
  /** The human friendly name of the object. */
  name?: Maybe<Scalars['String']>;
  /** Database id of the parent node */
  parentDatabaseId?: Maybe<Scalars['Int']>;
  /** The globally unique identifier of the parent node. */
  parentId?: Maybe<Scalars['ID']>;
  /** An alphanumeric identifier for the object unique to its type. */
  slug?: Maybe<Scalars['String']>;
  /** The name of the taxonomy that the object is associated with */
  taxonomyName?: Maybe<Scalars['String']>;
  /** The ID of the term group that this term object belongs to */
  termGroupId?: Maybe<Scalars['Int']>;
  /** The taxonomy ID that the object is associated with */
  termTaxonomyId?: Maybe<Scalars['Int']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};
// ...
export type MenuItemLinkable = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
  /** The globally unique ID for the object */
  id: Scalars['ID'];
  /** Whether the node is a Comment */
  isComment: Scalars['Boolean'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node represents the front page. */
  isFrontPage: Scalars['Boolean'];
  /** Whether  the node represents the blog page. */
  isPostsPage: Scalars['Boolean'];
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};
// ...
export type Node = {
  /** The globally unique ID for the object */
  id: Scalars['ID'];
};
// ...
export type UniformResourceIdentifiable = {
  /** The globally unique ID for the object */
  id: Scalars['ID'];
  /** Whether the node is a Comment */
  isComment: Scalars['Boolean'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node represents the front page. */
  isFrontPage: Scalars['Boolean'];
  /** Whether  the node represents the blog page. */
  isPostsPage: Scalars['Boolean'];
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};